Remove duplicate functions in climate/sensor.
Convert all data types correctly for climate.
